pyproject.toml
[project]
name = "cocad"
version = "0.1.0"
description = "COCAD - AI CAD Assistant"
requires-python = ">=3.10"
dependencies = [
    "groq",
    "python-dotenv",
]

[project.optional-dependencies]
dev = []
